reorganize store/write.ts, includes removing "quitGame" which was superfluous with setting the gameID to null
