+++ packages/webapp/src/store/write.ts
@@ -1,12 +1,12 @@
 /**
- * Actions that can modify the store.
+ * Function to modify the store, either directly or after transforming the input data.
  *
- * @module store/actions
+ * @module store/write
  */
 
 import { Address } from "src/chain"
 import * as store from "src/store/atoms"
-import { ErrorConfig, FetchedGameData, PrivateInfo } from "src/store/types"
+import { ErrorConfig, PrivateInfo } from "src/store/types"
 import { randomUint256 } from "src/utils/js-utils"
 
 import "src/utils/extensions"
@@ -17,62 +17,29 @@ const get = store.get
 const set = store.set
 
 // =================================================================================================
-// SET GAME ID
+// SIMPLE SETTERS
 
-/** Sets the game ID in the store. */
+/**
+ * Sets the game ID in the store.
+ */
 export function setGameID(gameID: bigint) {
   set(store.gameID, gameID)
 }
 
-// =================================================================================================
-// WAIT FOR UPDATE
+// -------------------------------------------------------------------------------------------------
 
 /**
- * Wait for the game data to be updated to the given block number (or beyond), resolving to the game
- * data. If the game is set to null at any point (indicating some kind of reset in the game state),
- * or if the timeout is reached, the promise resolves to null instead.
- *
- * Note that some operations (e.g. `drawInitialHand`) do not update the `lastBlockNum` field of the
- * game data and as such `waitForUpdate` is not suitable for use with these operations.
+ * Triggers the display a global UI error, or clears the error if `null` is passed.
  */
-export async function waitForUpdate(blockNum: bigint, timeout: number = 15)
-    : Promise<FetchedGameData|null> {
-
-  return new Promise<FetchedGameData|null>((resolve, _reject) => {
-
-    const unsubAndResolve = (result: FetchedGameData|null) => {
-      unsub()
-      resolve(result)
-    }
-
-    // Subscribe to the game data, resolve when receiving a state that satisfies blockNum req.
-    const unsub = store.store.sub(store.gameData, () => {
-      const gameData = store.get(store.gameData)
-      if (gameData === null || gameData.lastBlockNum >= blockNum)
-        unsubAndResolve(gameData)
-    })
-
-    // Maybe the game data is already up to date.
-    const gameData = store.get(store.gameData)
-    if (gameData !== null && gameData.lastBlockNum >= blockNum)
-      return unsubAndResolve(gameData)
-
-    // Initiate timeout.
-    setTimeout(() => unsubAndResolve(null), timeout * 1000)
-  })
-}
-
-// =================================================================================================
-// TRIGGER/CLEAR ERROR
-
-/** Triggers the display a global UI error, or clears the error if `null` is passed. */
 export function setError(error: ErrorConfig|null) {
   console.log(`setting error modal: ${JSON.stringify(error)}`)
   set(store.errorConfig, error)
 }
 
 // =================================================================================================
-// SET PRIVATE INFO
+// DERIVED SETTERS
+
+// -------------------------------------------------------------------------------------------------
 
 /**
  * Sets the private information specific to the given game and player in the preivate info store.
@@ -89,8 +56,7 @@ export function setPrivateInfo(gameID: bigint, player: Address, privateInfo: Pri
   })
 }
 
-// =================================================================================================
-// GET/INIT PRIVATE INFO
+// -------------------------------------------------------------------------------------------------
 
 /**
  * Returns the private information specific to the given game and player, initializing it if
@@ -119,19 +85,6 @@ export function getOrInitPrivateInfo(gameID: bigint, playerAddress: Address): Pr
 
   setPrivateInfo(gameID, playerAddress, privateInfo)
   return privateInfo
-}
-
-// =================================================================================================
-// QUIT GAME
-
-/**
- * Stop tracking the current game and reset all of its non-private data (the private data is kept
- * because it can't be recovered from the chain).
- */
-export function quitGame() {
-  set(store.gameID, null)
-  set(store.gameData, null)
-  set(store.hasVisitedBoard, false)
 }
 
 // =================================================================================================
